+++ lib/Models/Terria.ts
@@ -295,6 +295,11 @@ export default class Terria {
   @observable useNativeResolution = false;
 
   /**
+   * Gets ot sets whether to use antialiasing
+   */
+  @observable antiAliasingEnabled = false;
+
+  /**
    * Whether we think all references in the catalog have been loaded
    * @type {boolean}
    */