+++ CHANGES.md
@@ -34,6 +34,7 @@ Change Log
 * Only generate `tableStyles` for region columns if no other styles exist
 * TableAutomaticStylesStratum & CsvCatalogItem only returns unique `discreteTimes`s now
 * Specified specific terriajs config for ForkTsCheckerWebpackPlugin
+* Toggle in map settings to enable anti-aliasing of Cesium
 
 #### 8.0.0-alpha.43
 * Replace `@gov.au/page-alerts` dependency with our own warning box component. This removes all `pancake` processes which were sometimes problematic.